+++ Runtime/Levels/Level.cs
@@ -2,7 +2,9 @@ using UnityEngine;
 
 namespace Pixygon.Micro {
     public class Level : MonoBehaviour {
+        [ContextMenuItem("Get Pickups", "GatherPickups")]
         [SerializeField] private Coin[] _coins;
+        [ContextMenuItem("Get ActorSpawners", "GatherActors")]
         [SerializeField] private MicroActorSpawner[] _actors;
         [SerializeField] private Transform _playerSpawn;
         public Vector3 PlayerSpawn => _playerSpawn.position;
@@ -14,6 +16,14 @@ namespace Pixygon.Micro {
             foreach (var spawner in _actors) {
                 spawner.SpawnActor(loader);
             }
+        }
+
+        private void GatherPickups() {
+            _coins = GetComponentsInChildren<Coin>();
+        }
+
+        private void GatherActors() {
+            _actors = GetComponentsInChildren<MicroActorSpawner>();
         }
     }
 }