+++ Documentation/CHANGELOG.md
@@ -3,6 +3,7 @@
 ## Unreleased
 ### Added
 - First test of Pausing game
+- Automatic gathering of coins and spawners in level
 
 ## 0.5.6
 ### Fixed